Put action buttons into their own component
