+++ src/panel/PoiPanel.jsx
@@ -6,37 +6,21 @@ import Telemetry from '../libs/telemetry';
 import nconf from '@qwant/nconf-getter';
 import layouts from './layouts.js';
 
+import ActionButtons from 'src/panel/poi/ActionButtons';
 import PoiHeader from 'src/panel/poi/PoiHeader';
 import PoiTitleImage from 'src/panel/poi/PoiTitleImage';
 import OpeningHour from 'src/components/OpeningHour';
 import OsmContribution from 'src/components/OsmContribution';
 import PoiBlockContainer from './poi_bloc/PoiBlockContainer';
 
-import Store from '../adapters/store';
-import MasqFavoriteModal from '../modals/masq_favorite_modal';
 import CategoryService from '../adapters/category_service';
 import { openShareModal } from 'src/modals/ShareModal';
 
 import PoiStore from '../adapters/poi/poi_store';
 
-const store = new Store();
-const masqFavoriteModal = new MasqFavoriteModal();
-
-async function isPoiFavorite(poi) {
-  try {
-    const storePoi = await store.has(poi);
-    if (storePoi) {
-      return true;
-    }
-  } catch (e) {
-    return false;
-  }
-  return false;
-}
-
 export default class PoiPanel extends React.Component {
   static propTypes = {
-    poi: PropTypes.object,
+    poi: PropTypes.object.isRequired,
     isFromFavorite: PropTypes.bool,
     isFromCategory: PropTypes.bool,
     sourceCategory: PropTypes.string,
@@ -51,30 +35,10 @@ export default class PoiPanel extends React.Component {
     this.state = {
       isCollapsed: true,
       card: true,
-      showPhoneNumber: false,
-      poiIsInFavorite: this.poi.stored,
     };
     this.categories = CategoryService.getCategories();
     this.isDirectionActive = nconf.get().direction.enabled;
     this.isMasqEnabled = nconf.get().masq.enabled;
-
-    isPoiFavorite(this.poi).then(x => this.setState({ poiIsInFavorite: x }));
-
-    store.onToggleStore(async () => {
-      if (this.poi) {
-        this.setState({
-          poiIsInFavorite: await isPoiFavorite(this.poi),
-        });
-      }
-    });
-
-    store.eventTarget.addEventListener('poi_added', async () => {
-      if (this.poi && !this.state.poiIsInFavorite) {
-        this.setState({
-          poiIsInFavorite: await isPoiFavorite(this.poi),
-        });
-      }
-    });
 
     window.execOnMapLoaded(() => {
       const elem = document.querySelector('.poi_panel__content__card');
@@ -103,17 +67,6 @@ export default class PoiPanel extends React.Component {
     window.app.navigateTo('/favs');
   }
 
-  openDirection = () => {
-    if (!window.app) {
-      return;
-    }
-    window.app.navigateTo('/routes/', {
-      poi: this.poi,
-      isFromCategory: this.props.isFromCategory,
-      isFromFavorite: this.props.isFromFavorite,
-    });
-  }
-
   showDetail = () => {
     if (this.poi.meta && this.poi.meta.source) {
       Telemetry.add(Telemetry.POI_SEE_MORE, null, null,
@@ -138,31 +91,6 @@ export default class PoiPanel extends React.Component {
     fire('fit_map', this.poi, layouts.POI);
   }
 
-  toggleStorePoi = async () => {
-    if (this.poi.meta && this.poi.meta.source) {
-      Telemetry.add('favorite', 'poi', this.poi.meta.source);
-    }
-    if (this.state.poiIsInFavorite) {
-      await store.del(this.poi);
-      this.setState({
-        poiIsInFavorite: false,
-      });
-    } else {
-      if (this.isMasqEnabled) {
-        const isLoggedIn = await store.isLoggedIn();
-        if (!isLoggedIn) {
-          masqFavoriteModal.open();
-          await masqFavoriteModal.waitForClose();
-        }
-      }
-
-      await store.add(this.poi);
-      this.setState({
-        poiIsInFavorite: true,
-      });
-    }
-  }
-
   openShare = () => {
     if (this.poi.meta && this.poi.meta.source) {
       Telemetry.add('share', 'poi', this.poi.meta.source);
@@ -170,24 +98,6 @@ export default class PoiPanel extends React.Component {
     if (this.poi) {
       openShareModal(this.poi.toAbsoluteUrl());
     }
-  }
-
-  showPhone = () => {
-    const poi = this.poi;
-    if (poi && poi.meta && poi.meta.source) {
-      Telemetry.add('phone', 'poi', poi.meta.source,
-        Telemetry.buildInteractionData({
-          id: poi.id,
-          source: poi.meta.source,
-          template: 'single',
-          zone: 'detail',
-          element: 'phone',
-        })
-      );
-    }
-    this.setState({
-      showPhoneNumber: true,
-    });
   }
 
   closeAction = () => {
@@ -217,29 +127,6 @@ export default class PoiPanel extends React.Component {
     this.setState({
       card: true,
     });
-  }
-
-  shouldPhoneBeHidden() {
-    return this.poi &&
-      this.poi.isFromPagesjaunes &&
-      this.poi.isFromPagesjaunes() &&
-      !this.state.showPhoneNumber;
-  }
-
-  renderPhone() {
-    if (this.shouldPhoneBeHidden()) {
-      return <button className="poi_panel__action icon-icon_phone poi_phone_container_hidden"
-        onClick={this.showPhone}
-      >
-        <div>{_('SHOW NUMBER', 'poi')}</div>
-      </button>;
-    }
-    return <a className="poi_panel__action icon-icon_phone poi_phone_container_revealed"
-      rel="noopener noreferrer external"
-      href={this.poi.blocksByType.phone.url}
-    >
-      <div>{this.poi.blocksByType.phone.local_format}</div>
-    </a>;
   }
 
   render() {
@@ -325,30 +212,12 @@ export default class PoiPanel extends React.Component {
               <PoiHeader poi={this.poi} />
               <PoiTitleImage poi={this.poi} iconOnly={false} />
             </div>
-            <div className="poi_panel__actions">
-              <button className={classnames(
-                'poi_panel__action',
-                'poi_panel__actions__icon__store',
-                {
-                  'icon-icon_star-filled': this.state.poiIsInFavorite,
-                  'icon-icon_star': !this.state.poiIsInFavorite,
-                })
-              }
-              onClick={this.toggleStorePoi}
-              >
-                <div>{this.state.poiIsInFavorite ? _('SAVED', 'poi') : _('FAVORITES', 'poi')}</div>
-              </button>
-              <button className="poi_panel__action icon-share-2" onClick={this.openShare}>
-                <div>{_('SHARE', 'poi')}</div>
-              </button>
-              {this.isDirectionActive &&
-                <button className="poi_panel__action icon-corner-up-right"
-                  onClick={this.openDirection}>
-                  <div>{_('DIRECTIONS', 'poi')}</div>
-                </button>
-              }
-              {this.poi.blocksByType && this.poi.blocksByType.phone && this.renderPhone()}
-            </div>
+            <ActionButtons
+              poi={this.props.poi}
+              isFromCategory={this.props.isFromCategory}
+              isFromFavorite={this.props.isFromFavorite}
+              isDirectionActive={this.isDirectionActive}
+            />
             {this.poi && this.poi.id.match(/latlon:/) && this.categories &&
               <div className="service_panel__categories--poi">
                 <h3 className="service_panel__categories_title">
